requirements.txt
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
ecdsa==0.14.1
fastapi==0.63.0
h11==0.12.0
httptools==0.1.1
jwcrypto==0.8
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.3
python-dotenv==0.15.0
python-jose==3.2.0
python-multipart==0.0.5
PyYAML==5.4.1
rsa==4.7
six==1.15.0
starlette==0.13.6
uvicorn==0.13.3
uvloop==0.14.0
watchgod==0.6
websockets==8.1
appdirs==1.4.4
bcrypt==3.2.0
black==20.8b1
certifi==2020.12.5
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
ecdsa==0.14.1
fastapi==0.63.0
h11==0.12.0
httpcore==0.12.2
httptools==0.1.1
httpx==0.16.1
idna==3.1
jwcrypto==0.8
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.8.1
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.3
python-dotenv==0.15.0
python-jose==3.2.0
python-multipart==0.0.5
PyYAML==5.4.1
regex==2020.11.13
rfc3986==1.4.0
rsa==4.7
six==1.15.0
sniffio==1.2.0
starlette==0.13.6
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
uvicorn==0.13.3
uvloop==0.14.0
watchgod==0.6
websockets==8.1
